+++ apps/native-component-list/src/screens/UI/JetpackComposePrimitivesScreen.android.tsx
@@ -1,4 +1,4 @@
-import { Container, Column, Row, Text } from '@expo/ui/jetpack-compose';
+import { Container, Column, Row, Text } from '@expo/ui/jetpack-compose-primitives';
 import { useState } from 'react';
 import { Button, View, StyleSheet, Text as RNText } from 'react-native';
