+++ packages/expo-ui/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### 🎉 New features
 
 - [android] Add shape component. ([#36964](https://github.com/expo/expo/pull/36964) by [@aleqsio](https://github.com/aleqsio))
-- Added low-level `jetpack-compose` export. ([#36257](https://github.com/expo/expo/pull/36257) by [@kudo](https://github.com/kudo))
+- Added `jetpack-compose-primitives`. ([#36257](https://github.com/expo/expo/pull/36257) by [@kudo](https://github.com/kudo))
 
 ### 🐛 Bug fixes
