+++ packages/expo-ui/jetpack-compose-primitives.d.ts
@@ -1,0 +1,1 @@
+export * from './build/jetpack-compose-primitives';
