--- packages/expo-ui/jetpack-compose.ts
@@ -1,1 +1,0 @@
-export * from './src/JetpackCompose';
